Use the load_dtrace_modules from the build tree when doing uninstalled tests.
This is two-part: firstly, install load_dtrace_modules into the uninstalled
build/dlibs directory when testing, and secondly, look for it in that
directory (the syslibdir, during uninstalled testing) rather than only looking
in the DTRACE_LIBDIR hardwired at compile time.

This means we have to construct the directory to look for load_dtrace_modules in
at runtime.  Rather than enter a hell of potential overruns -- since the length
of environment variable values is nearly unbounded, and filenames are not -- we
do it with the GNU-specific asprintf().

Orabug: 19821254