Use the new module autoloading mechanism to load modules for the testsuite.
This means we can keep the module autoloading mechanism from rotting untested
and rip out the specialized module loading code from the testsuite... and
replace it with more code, since we have to not only load the modules but verify
that they loaded, since the tester will no longer get any sort of error message
if the loading fails.  While we're at it, make failure to load fatal: it was
only ever nonfatal because I wasn't sure module loading at test time would work,
and it does, and if a module fails to load disaster is certain in any case.

We do this by using DTRACE_MODULES_CONF to point dtrace at the test/modules
file, which it then reads as if it were /etc/dtrace-modules and loads all the
modules therein.  (It only does this if dtrace.ko isn't loaded, but runtest.sh
ensures that happens just before.)

We also have to set up the DTRACE_OPT_SYSLIBDIR earlier than before, so that it
is in place when module loading happens, so that dtrace has a chance to find
load_dtrace_modules in the uninstalled-testing case.

The special @@ tags, @@unload-only and @@quiet, used in test/modules, are not
understood by load_dtrace_modules, so we turn them into ##unload-only and
##quiet, which are valid comments and thus ignored.

Orabug: 19821254