No longer use fts() to find kernel module name->path mappings.
This is deprecated, doesn't work with large-file APIs, and is generally
considered bad.  Of course, its replaement -- nftw() -- has other problems,
notably the 'qsort() bug' that it calls a function of your choice to do the work
of iteration but provides no way to pass an arbitrary pointer to it, so you
can't get any state in (like the dtp pointer).

But of course we don't need to update kernel module locations often: we just
want to protect against some hypothetical multithreaded libdtrace consumer
calling dt_module_lookup_by_name() for the first time simultaneously in multiple
threads, and then we can use a static global to pass the necessary state.  So
add a static mutex to do that.

This also lets us fix up the memory allocation used by the module-iteration
path, avoiding a static MAXPATH-sized buffer for the module path in favour of
the dynamic allocation we're doing everywhere else anyway.