Never, ever exit on OOM.
Normally, exiting on OOM is a good strategy for programs: it avoids introducing
almost-impossible-to-test error cases.  However, debuggers and tracers that use
ptrace() to introduce breakpoints on traced processes cannot ever do this: if
they exit unexpectedly, their victims are left with breakpoints outstanding in
them and will be killed as soon as the breakpoint hits them.

In the absence of a better approach, simply fix a bunch of places where we exit
unexpectedly.

This is not a complete fix: running out of stack space will still OOM us, as
will bugs.  In the future we should either arrange for the kernel to roll back
our breakpoints, or create a child process which the dtrace userspace code
informs about all outstanding breakpoints and whose sole purpose is to poke the
right text back in as fast as possible once the parent dies.  That's racy and
ugly though: a proper kernel solution, perhaps involving new
ptrace(PTRACE_POKETEXT_ON_DETACH) and ptrace(PTRACE_DETACH_ON_TRACEREXIT)
requests, is probably better.

Orabug: 20014606