Improve .d.in per-kernel script substitution.
It turns out that defining symbols named things like KERNEL_3.18 is distinctly
inconvenient: it leads to massive repetition in the D file if you want to do
something like replace a string in some set of kernel versions.

So this change introduces a new macro, define_for_kernel, which can be
used to say things like "replace this string with this string for all these
kernel versions".  The "string" is an M4 macro definition, so can use M4
parameters to introduce sort of fake accessors to access things that vary
between kernels in a generic way.

The syntax is somewhat baroque, as is typical of M4:

define_for_kernel([[version]]..., [[string]], [[result]], ([[otherwise]]))
expands to a macro that replaces [[string]] with [[result]], but only on kernels
in the quoted list [[version]]; if [[otherwise]] is provided, other kernels get
this expansion instead.

So you could say

define_for_kernel([[[[3.8]],[[3.18]]]],[[foo bar]],[[foo baz]])

and the string 'foo bar' would be replaced with 'foo baz' on kernels 3.8 and
3.18 only, and would be left alone otherwise. Or you could say

define_for_kernel([[[[3.18]],[[3.19]]]],[[ugid]],[[$1.val]],[[$1]])m4_dnl

and the string 'ugid(anything)' would be replaced with 'anything.val' on 3.18
and 3.19, and 'anything' on other kernels.

Note: the [[version]] list is a list of *quoted* versions, so each version is
quoted twice!

This is somewhat ugly because of the terrible quoting, but is a lot less ugly
than naming everything over and over again.

As part of this, we pass in a single M4 macro, KERNEL, which expands to the
kernel version being processed, rather than passing in a macro whose name
represents the kernel version.  We also split up the M4 scripts more, and pull
in a couple of scripts to implement advanced concepts like, uh, loops from the
M4 distribution's examples tree (licensed with 'unlimited permission to copy
and/or distribute it, with or without modifications, as long as this notice is
preserved'; this is not distributed in any case but is only part of the build
system).

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>