Fix preprocessing of #!-prepended D scripts.
The code in DTrace to handle this case (by seeking past the #! line, up to the
linefeed, then handing that fd to its preprocessor child via fd inheritance and
/dev/fd) was broken in multiple ways:

 - it relied on fseeko() moving the file descriptor of the underlying Unix file
   to the same place, which is not at all guaranteed due to buffering (and glibc
   seeks before the requested point and then prereads to a point after it to
   fill the stdio buffer)

 - it relied on the preprocessor's open of /dev/fd/* leaving the file position
   unchanged, which is unfortunate because open() is defined to start at the
   beginning or end of the file and *cannot* leave the file position unchanged:
   open() of /dev/fd/* does *not* act just like dup() in this respect at least

But this is all irrelevant because the whole thing was hopeless: GCC's
integrated preprocessor checks to see if it's been opened on a regular file, and
counts the number of bytes it reads and warns if they're not equal to the size
returned from stat().  So even if this technique had worked, it would elicit
a warning on every preprocessor invocation.  Not good.

Instead, iff the preprocessor is in use and this is a seekable file (the same
conditions as before), and there is a #! present, we fork off another child,
feeding it the input file over stdin, whose sole purpose is to splice() its
stdin to its stdout.  Then we feed that stdout into cpp's stdin. i.e. the
equivalent of converting

cpp /dev/fd/4 /dev/fd/5 <4 input.d 5> (dtrace parent)

into

cat < input.d | cpp /dev/fd/5 5> (dtrace parent)

only more efficient (and, in both cases, not running from the start of input.d
but from the end of the first line).  This converts the seekable file we were
handing cpp into a nonseekable pipe, and quashes its warning message.

This requires a little adjustment elsewhere: dt_pragma_line() knew about the
/dev/fd/* trick, erasing that string from the #line's emitted by the
preprocessor and translating them into the name of the main D file, so error
messages referred to the right file.  But this trick doesn't work on errors from
the preprocessor itself: some expected test results needed adjustment to turn
/dev/fd/4 into /dev/stdin (which is hardly ideal output, but is much better than
/dev/fd/4, I think).

Orabug: 20746827