Per-kernel D libraries, specfile portion.
The old kver38 definition is turned into a Lua array of kernel names
(as seen in RPM version numbers, uname -m, and /usr/src/kernels).  This is
then used to emit appropriate build requirements and something that is
subjected to shell filename expansion to yield the names of the
KERNELS= from the names of directories under /usr/src/kernels.