Fix building with DESTDIR set.
The DESTDIR makefile variable is only supposed to affect installation, but
because we applied it to LIBDIR, BINDIR etc, and then used those variables in -D
arguments to the compiler, it affected the generated binary if set at make time.

Fix this by splitting the DIR variables into names of the
form *DIR (e.g. LIBDIR) which do not have DESTDIR applied, and INST*DIR
(e.g. INSTLIBDIR), which do, and change all installation targets to use
the INST* variables exclusively.