Fix an errno niggle.
In Pbkpt_continue(), we were looking at an errno result from ptrace(), but quite
possibly after other things (e.g. kill()) had run and perhaps perturbed it.  The
check was for a can-never-happen condition, but this is still a bug (among other
things, it's not good if a can-never-happen condition is believed to have
occurred when it in fact has not.)