Transition to an explicit stack for Ptrace()/Puntrace() states.
When you Puntrace() a process you have previously Ptrace()d, it is important to
return to the same run state that the process was in at that point (if it was
running before you Ptrace()d, it should be running now, if it was stopped, it
should be stopped now, etc).  The model used for Ptrace() and Puntrace() was to
have Ptrace() return the state of the process before invocation, and having
Puntrace() take it: thus, we stored those states on the C stack.

This all falls apart when you have multiple threads that may be in distinct run
states, particularly if (as we do) you want different forms of call to trace all
threads in the process or only one.  We have to transition to an explicit
Ptrace()/Puntrace() stack so that we can later move this into the per-thread
structure and have everything just work.