Add CPU hotplug support; rip out include/sys/processor.h, and all that it implies.
So dtrace had this type processorid_t.  I'd just assumed it was the ID of a CPU,
and its predominant use in <dtrace.h> suggested as much: it's used where a CPU
number is desired.

But then Wim reported that you can't actually include <dtrace.h> from outside
the DTrace source tree, because it pulls in <sys/processor.h> which is not
installed.  Since this makes it impossible to write DTrace consumers this is
distinctly suboptimal.

This whole area of DTrace shows signs of incomplete porting: we have
processorid_t which is simply a typedef for 'unsigned int' and which is
apparently meant to denote a CPU ID, and a bunch of #defines which we mostly
don't use: the only one we use is P_STATUS, which the horrendously misnamed
dt_status() in dt_subr.c passes to p_online() as the 'new_status' parameter to
see if a CPU is online.  p_online() then ignores it and returns the hardwired
constant 2 (apparently intended to be P_ONLINE), which is passed back to
dt_status() to its only caller, which then ignores it and just considers -1 to
mean 'not online'.

So rip this whole mess out and declare p_online() to be a function that takes a
CPU ID (and no bizarre 'new_status') and which simply returns -1 to indicate
that a CPU is offline and 0 otherwise, by consulting
/sys/devices/cpu/cpu$num/online (if not present, the CPU is online).

To flag up this, we rename dt_status() to dt_cpu_status() (since it has nothing
to do with all the other uses of the word 'status' around DTrace and is
specifically testing a *CPU's* status), and rename the dtv_status() member in
the public dtrace_vector_t type to dtv_cpu_status() to point out to any users
outside DTrace that this has changed.  (Not that there *can* be any users,
because any users would have had to #include this header, which would fail...)

Orabug: 21154502