Handle non-breakpoint-related SIGTRAPs in child processes again.
This was turned off in commit e7d3c448c22ae5470115abe5722f705156987dde because
we were being spuriously hit by SIGTRAPs in multithreaded processes.  Now we
remove all things that might drop breakpoints in such processes, this is no
longer appropriate, and we can reinject them, after verifying that they are not
in fact breakpoint-related.  (This also means that we can re-enable the test for
proper signal handling, which was being thrown off by the discarding of
SIGTRAP.)