Ensure that we are PS_TRACESTOPped whenever ptrace() implicitly stops us.
There are several tracing operations (TRACECLONE, TRACEEXEC etc) which
implicitly stop the process.  If we don't track this in the process's
ps_prochandle->state, and then do something like removing breakpoints which can
call Pwait(), we can find ourselves attempting to interrupt a process that's
already stopped, then waiting forever for the interrupt to be acknowledged.

Fix by marking the process as halted in trace state while processing clone,
fork, vfork and (just in case) exec events.  For everything but the exec()
(which throws away and recreates all our state tracking eventually, including
the ->state field), reset it afterwards, when we resume the stopped process.