Add a 'make install-test' rule, producing an installed copy of the testsuite.
The installed testsuite looks much like a top-level dtrace source tree,
only without any of the source code:

-rw-r--r--  1 root root  1297 Jun 30 10:43 GNUmakefile
-rw-r--r--  1 root root  5762 Jun 30 10:43 Makecheck
-rw-r--r--  1 root root  1168 Jun 30 10:43 Makeoptions
-rwxr-xr-x  1 root root 43324 Jun 30 10:43 runtest.sh
drwxr-sr-x 11 root root  4096 Jun 30 10:43 test

By default this is located under the dtrace syslibdir, in
/usr/lib64/dtrace/testsuite.  (It does not belong under /usr/share because
its test/triggers directory contains architecture-dependent code).  You
can run the testsuite by cd'ing in here and doing a make check, or any
other testing target, just as in the source tree.  All the targets do
installed tests, against the installed dtrace.  (You can still call
runtest.sh directly, without passing --use-installed, if you really think
you need to.)

The GNUmakefile is derived from Makecheck.installed.in in the source tree,
sedded to provide an appropriate definition of TESTDIR.  It defines a
variable to indicate to Makecheck that all testing is installed-dtrace
testing, and a bunch of makefile targets to suppress the usual automatic
trigger generation that happens before running installed tests.  We also
install a variant of Makeoptions that describes the much more limited
options available to the installed testsuite.  Makecheck comes straight
from the source tree and is only changed insofar as is needed to allow
various paths (e.g. to runtest.sh, and to the git version file) to change.

There is one other difference: doing a 'make check-installed' does not
run the internals tests (it cannot, because that would require building
dtrace, and the whole point of 'make check-installed' is to avoid doing
that, and to test the locally installed copy).  An installed testsuite
does not have this restriction (since, if it is installed, it will be
installed from an RPM that was generated from the same source tree as
the installed dtrace), so an installed testsuite *can* run the internals
testsuite.

We modify runtest.sh in only one tiny way: we can no longer rely on the presence
of the Makefunctions file to discern the root of the testing tree (and thus a
point where upward searches for test.options should terminate).  We can rely on
Makecheck instead, which is a better idea anyway since it is more clearly
testsuite-related.