Improve dependencies in version-reporting machinery.
There are two faults here (neither especially serious, but both annoying).

First, having .git-version depend on nothing is problematic: if you do a commit,
you should see .git-version updated, but it isn't.  So have it depend on the git
index, so that any index update will force a regeneration of .git-version.  But
that would break things if this was a tarball checkout, so have a dummy target
for .git/index in that case.

This shows up another problem.  While it is tolerable for cmd/dtrace.c, having
dt_git_version.h in the _SRCDEPS of libdtrace-build is inefficient: this puts it
in the dependency list of *everything* in libdtrace, even though only one file
depends on it, and forces a complete rebuild of libdtrace whenever a git commit
is done.  Manually specify the appropriate dependency instead.