Non-verbose testing should still emit the test name into the logfiles.
This wasn't happening because the code to emit the test name was
the same as the code that printed the test name to stdout, so when
the latter was suppressed, so was the former.  Logfiles shold be
unaffected by quietness state.

As a side benefit, this lets us strip out the ugly scoped-force-out
machinery (controlled by an environment variable) in favour of a
new force_out_only(), which prints to stdout but *not* the log
or sumfile, and is used exclusively to print out the test name
(and only the test name) on failure.  It also de-tangles pass()
and fail() noticeably.

Orabug: 21769905