Revamp 'make check-module-loading'.
This piece of the testsuite is an annoying piece of code duplication
which has now been the source of multiple bugs: there is code to load
the modules both inside runtest.sh for 'make check' and inside Makecheck
for 'make check-module-loading', and they are not quite the same.  (The
latest fault is that the Makecheck version doesn't autoinstall modules
if they're not there.)

So rip out the Makecheck version entirely, and add a new ./runtest.sh
option --load-modules-only which causes runtest.sh to exit after
silently unloading and then non-silently reloading the modules.

Add a new target 'make check-installed-module-loading' to indicate
that the installed DTrace should be used to invoke module
autoinstallation if needed.