Fix ELF object base address computation in the absence of rtld_db.
If rtld_db cannot initialize, we have to determine the library base
address (the corrective offset which is applied to names in the symbol
table to yield the corresponding addresses in memory) ourselves.

The method we were using to do this before was crazy and bore no
resemblance to anything in the ELF specification (nor to the entirely
different and zany method used on Solaris, which works only because of
implementation details of their dynamic loader).  It also had piles of
magic hackery in so that we at least got the right answer for ld.so's
library base address, but the method we were using for that was nuts and
involved a name lookup.  We now do precisely what the ELF specification
says: find the PT_LOAD segment with the lowest p_vaddr and subtract that
from the load virtual address (derived ultimately from /proc/$pid/maps).
This unfortunately means we have to iterate through all the phdrs even
though 99.999% of the time the PT_LOAD segment we want is the first one
-- so it's a good thing we only have to do this sometimes.

To be specific, we do it when rtld_db isn't working (which includes when
we're noninvasively tracing or when rtld_db can't figure out how to dig
into the version of glibc under consideration) and also where a non-PIE
main executable or ld.so is concerned, because ld.so doesn't bother
keeping its own base address up to date in its own data structures when
it relocates itself, and the main executable isn't a shared library so
none of this applies.

Orabug: 22106441