Fix termination condition when searching for test options to extract.
Both runtest.sh's exist_options and extract_options functions walk up
the filesystem hierarchy searching for options to extract.  This means
they have to know when to stop searching.  The termination condition for
option saerching used to be the existensce of the Makefunctions file,
but since this doesn't exist in testsuite RPMs this was adjusted to be
Makecheck in c382355.  Unfortunately this was only adjusted in
exist_options(), so it was possible for a caller which called
extract_options() with nonexisting options to loop forever.  This never
happens modulo other bugs, so this problem was overlooked.