Change trigger mechanism to handle excessive startup time.
Due to the difference in startup times for DTrace on x86 and sparc,
and the overall uncertainty of startup timing due to other system
load, the trigger in the tst.modsdt.d testcase could fire too soon,
causing the test to fail because the trigger ended prior to probes
having been enabled.

This commit starts the trigger from a BEGIN action, ensuring that it
will not start prior to DTrace having completed its startup.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>