Harden tests that make tempdirs against multiple invocations.
There are numerous tests that make temporary directories under $tmpdir.
These assume that because $tmpdir is recreated on every runtest
invocation that they cannot collide with any existing directory.

Unfortunately, because you can specify a single test more than one in a
runtest line, this assumption is wrong, and when they collide, mkdir
emits an unexpected error message and the test spuriously fails.

Use $$, $RANDOM, and mkdir -p to make sure this problem goes away.
(We still don't remove the directory because it can be useful
for problem diagnosis, and it's removed at the end of runtest execution
by default anyway.)