Avoid lock-count information loss leading to shutdown asserts and hangs
Because POSIX recursive mutexes do not work in conjunction with
condition variables, dtrace has long maintained its own lock count for
the dpr_lock used to communicate state changes (stopped process, dying
control thread etc) between the main thread and a given process-control
thread that holds the ptrace on a process.

Unfortunately, our handling of this across condvar operations
(proxy_call()s, dtrace shutdowns, etc) was less than ideal.  We were
taking the lock (as you are obliged to before waiting on a condvar),
remembering the old lock count and zeroing it, then waiting on the
condvar: then, on return, we put the old lock count back in.

This works fine *if* the call to the process-control thread always takes
and releases the lock the same number of times, since the lock count
will then always be zero on exit, just as it was on entry.
Unfortunately, this is not always true: there are several operations
(such as process grabbing and creation) that grab the lock (usually
because it initiated a ptrace() and left the process halted for later
signalling) and then leave it grabbed, and there are also several
operations that pop the lock (such as process death, reversing the grab
just mentioned).  The key thing about these grabs is that they are
long-term: they can span across a proxy_call() and last for a while.

(This sort of long-term lock holding might seem risky, but this
long-term lock hold only happens between process grab/startup and
dtrace_proc_continue(), and because the process is known stopped at this
point, the dtrace main thread never tries to do anything with it or take
the dpr lock, except as needed to make condvar calls to the control
thread.)

So rather than brushing this under the carpet, properly describe it by
splitting the lock count into two, one for the control thread and one
for the main thread (all other threads, really, but the only other
thread that touches dpr locks is the main thread).  Then we can
increment and decrement the right count automatically, and can throw
away the stack-based attempt to remember the old lock count around proxy
calls.

We also have to adjust shutdown slightly.  If Prelease() sees that the
ptrace_count is nonzero, it will release the ptrace() and do a single
unlock to counteract the long-term lock done by Ptrace() (just as
Puntrace() does).  So the Prelease() call in dt_proc_control_cleanup()
can sometimes release the lock, but if there were multiple nested levels
of Ptrace() it will not.  So check for this case, and reduce the lock
count far enough that another unlock will truly release the lock: then
release it.  (If there are other nested Ptrace() calls in force, they'll
all return PS_DEAD at the next operation and unwind because the process
is gone: there is no longer any state meaningfully present for a lock to
protect).  Because this means we're moving the lock up, move up the
dpr_done assignment that the lock is guarding too, and shift the
broadcast on the condvar into the unlock's old place at the end of the
function, to prevent dtrace from finishing shutdown and freeing the dpr
structures before the thread cleanup that relies on those structures is
complete.

We also add a test for this case (in the expensive testsuite because it
takes ages for this bug to show itself sometimes).

Orabug: 22824594