Improve tests of aggregation clearing, avoiding spurious failures
Using clear() in the middle of D execution from a profile timer appears
to sometimes cause the last accumulations into the agg to be lost if it
happened at the same instant, probably because of cross-CPU coordination
issues.  These are both not a bug and not what the tests in
test/unittest/aggs/tst.clear* are meant to be testing: rather, they're
supposed to be verifying that if you clear an agg, it's clear, and that
once you fill it up again, it fills up without incident, just as if
DTrace were newly started.

So add some result postprocessors that validate this.  (If you still need
to see the raw output, just move the postprocessors out of the way.)

Orabug: 22024070