proc: fix overrun / type clash with PTRACE_GETEVENTMSG
When processes attached to DTrace via -c fork() or clone(), it has to
learn about it in order to remove associated breakpoints. This is done
via PTRACE_EVENT_{FORK,VFORK,CLONE}, all of which do a
PTRACE_GETEVENTMSG to get the PID of the new child, remove breakpoints
init, then resume it. Unfortunately, ptrace()'s interface for doing this
is somewhat baroque: it returns the PID in the variable pointed to by
the data argument: although this is nowhere documented, this variable
is considered to be an unsigned long, and on platforms like SPARC
where sizeof (unsigned long) > sizeof (pid_t), if you use a pid_t like
you'd expect to be able to for PIDs this will overrun the variable and
may cause an -EFAULT, leading to DTrace losing track of the child and
its staying frozen at a tracing stop until DTrace terminates.

Fix by using unsigned longs in appropriate places.  (If there are
platforms on which sizeof (pid_t) > sizeof (unsigned long), this
interface will fail completely, but it seems most unlikely that Linux
could ever be made to work on any such architectures.)

Orabug: 24447730