Don't call the dtrace prochandler for internally-grabbed processes.
Currently, if dtrace grabs a process due to a ustack() or other internal
feedback (i.e., not as a result of -c or -p from the command line) and
that process dies while it is grabbed, we call the process death
notifier for it.  This is wrong, since it leads to the caller receiving
notifications of the death of processes it never asked to be grabbed and
probably does not know about.

Fix by adding a new internal-use-only DTRACE_PROC_NOTIFIABLE flag and
using it to mark all grabs initiated from the external API: processes
grabbed without this flag do not call the prochandler.

Orabug: 24829169