dis: dump the offset as well as the instruction counter
The OFF column in DIF disassembly is actually not the offset but
a simple instruction counter, leading to endless confusion when trying
to match it up with the offset returned in kernel error messages (which
is a true offset).  Dump the offset too.

Also boost the column width of both columns to three digits, since it
is fairly routine to have predicates > 100 bytes long.