Add test case for bug 26089286
It was determined that FBT return probes on x86_64 are executing in a
context where in_irq() returns true, even though the code that triggered
the probe is not executing in IRQ context.  That means that TLS variables
(self->var) in D actions will not hash to the same key for FBT entry
probes vs FBT return probes.

These tests verify the fix for this problem by verifying the value stored
to a TLS variable at entry probe time against the value retrieved for
that TS variable at return probe time..

Orabug: 26089286
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>