Move time consuming tests to the test/stress hierarchy
Two tests in the dtrace-util directory (DestructWithModule and
ZeroModuleProbes) depend on using *all* FBT probes which right
now causes a significant slowdown.  The same happens with the
safety tests.  Because they need a long timeout to be able to
determine whether they pass, they are being moved to the stress
hierarchy of tests.