Look for cpp and ld along the PATH.
The default linker and preprocessor used internally by DTrace are hardwired
to /usr/bin/ld and /usr/bin/cpp. This is unwise, since it is commonplace to
replace the compiler by adjusting $PATH, and the compiler driver (for which
cpp is a frontend) looks for its own components along the path as well. This
means that if you e.g. scl enable devtoolset-6, you end up with a GCC 4.4 cpp
trying to invoke pieces of GCC 6. Disaster results.

The solution is to remove the path from the linker and assembler and use path
searches here, too, just like the compiler itself does.

Orabug: 26396530