tests: deadman timer detection: use it in the stress testsuite
Now we have a working deadman timer, we have to support it properly:
it's often a feature when it fires, not a bug, since the alternative
is probably a dead system or a watchdog timer reboot.

Move all tests that are expected to trip the deadman (roughly, all
those referencing fbt:::) into the stress testsuite, and give the
entire stress testsuite both a @@timeout of 600s, @@timeout-success
(so that timeouts, when they happen, are considered XFAILs) and the
new @@deadman-success option, which is like @@timeout-success except
that it detects deadman firings and considers them XFAILs too.

(This may prove inadequate: if a test fires the deadman yet still
passes, this will cause spurious XPASSes and we'll have to complicate
things some more. But it's a good start.)