Prevent dpr destruction proceeding before control threads are cleaned up
There are yet more problems around process termination.  To wit, after
the fix for bug 22824594 in commit d1bfc61e24b9, we started to release
the dpr_lock earlier in the cleanup phase, because Prelease() could
sometimes release the lock in any case and it was a long way further
up the function than the explicit unlock removed in that commit.

Unfortunately, this means the dpr_lock is released before we start
closing proxy filehandles, enqueueing death notifications,
releasing proxy calls etc, and all of this requires access to the dpr;
but if we have already released the dpr_lock, the main thread could
be blocked on the lock acquisition in dt_ps_proc_destroy(): as soon
as that is released, we'll note that the control thread is "dead"
(because we set dpr_tid to zero on that final unlock) and free
the dpr, following which all the references in the still-alive
cleanup thread are traversing a freed pointer.  Chaos can result.

To fix this, migrate the entire release/unlock dance much further
down cleanup, doing it at the very end: no further dpr dereferences
happen later, with us going so far as to cache the value of dpr->dpr_pid
used in a debug message in a local variable to avoid one.

As a side-effect this means we can simplify out some work around
death-notification queuing because it's easier to tell if we grabbed a
process successfully if we haven't killed it yet.

This change is safe because we are widening a locking region, not
narrowing one, and none of the new work done in that widened region
requires the main thread to be responsive. (Prelease() can do proxy
calls if it is called from the main thread, but when called from the
process-control thread, as here, it never does such a call because
that would always lead to a deadlock.)

In related work, the 'have we exec()ed' check in proxy_call() is unsafe
because it is checking to see if we have exec()ed but not if we have
quit immediately afterwards, leaving us with no control thread; and
Prelease()'s process-is-dead path was not unlocking any outstanding
locks, leaving us with a permanently leaked lock.

Finally, ditch an obsolete debugging message: if a proxy call discovers
that the control thread is dead, this is a perfectly normal consequence
of process termination: don't emit a debug message, because we might
emit scads of them, and do set errno appropriately.

Orabug: 26528776
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>