UPL relicensing, stage 1: manual modifications.
This relicenses everything that cannot be relicensed automatically
and is long enough to be meaningfully copyrightable at all (and
not automatically generated and uncopyrightable, like the .r files).

In addition, various copyrights that span lines or are sufficiently
nonstandard that the automated tooling will not pick it up are
reworded to the Oracle standard form.