Allow overriding of kernel list from command line
You can now use --define option for rpmbuild to override kernels to be
used from commandline so there is no need to edit specfiles.  When used
this way it is required to provide full kernel name (no suffixes are
appended in this case).  This way you are able to use development kernel
packages produced by internal autobuilders.

Signed-off-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>