New packaging scheme for user-space
With DTrace as a part of the kernel source tree we need to adjust our
packaging scripts.  It is not possible to use kernel-uek-headers for
delivery of our header files so we need to ship them on our own.  Header
files are GPL licensed where user space components are UPL licensed so
new package has been create to ensure we don't ship mixture of both in
single package.

RPM packages:

    - dtrace-utils ... required for DTrace users (as before).
    - dtrace-utils-devel ... ships kernel headers for developers.
                        (Obsoletes dtrace-modules-shared-headers)

New capabilities have been introduced:

    - dtrace-kernel-headers ... signals version of headers available
                                from kernel sources.
    - dtrace-headers ... signals version of headers shipped by the
                         dtrace-utils-devel pacakage.

New macros for rpmbuild:

    - local_build ... Use when you want to build RPMs and use headers
                      localy installed in /usr/include/linux/dtrace
    - build_kernel ... Use headers from matching kernel-devel pacakge

To use DTrace you need:
	yum install dtrace-utils

To develop DTrace you need:
	yum install dtrace-utils-devel

Signed-off-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>