CTF decoupling support.
Arrange to load CTF from the CTF archive if present.

To do this, we repurpose the already-existing dt_kernel_module_find(),
which used to just dig up the path to the kernel module (populating all
the paths if necessary) to first check for a valid CTF archive in the
dt_module_path.  Right now, the name of this archive is hardwired to
'vmlinux.ctfa', but in future we may change things to load and search a
list of CTF archives to support KSplice and similar dynamic patching
mechanisms.  If we find it, we load the shared CTF from it, then try to
load the CTF for the requested module: if either fails, we fall back to
the old machinery and hunt the module down.  This means that we still
work with old kernels: in particular, we still support the old format of
ctf.ko, and will until UEK4 is obsoleted.  Once opened, the archive
remains open until dtrace_close() time: the archive is almost entirely
(privately) mmap()ed, and only one page of that mapping is written, so
the amortized memory hit of even a huge CTF file is very low.

A new DT_DM module flag, DT_DM_CTF_ARCHIVED, is added to indicate that a
module's CTF file is loaded from the archive.  We suppress loading of
module section data if this flag is on, and also decompression: see
below.

Modules in the archive have another difference from the old regime:
rather than being pulled in as a blob and manually decompressed, we use
the built-in support in libdtrace-ctf for compressed modules and rely on
it to do all uncompression.  (This is necessary in any case because the
archiving code has to be able to look up the size of the module from the
CTF file header, so that header must be uncompressed: but we exploit
this to gain additional performance by compressing only big CTF files.
Right now, that limit is arbitrarily set to one page: 85% of CTF files
fit in under this limit -- though most are probably very rarely accessed
-- adding only a few dozen kilobytes to the size of the installed
kernel, and reducing memory consumption somewhat because these files can
be read directly from their memory mapping.  Compressing large CTF files
will probably always be worthwhile, because opening a CTF file means
reading the whole thing in, and disk I/O is slow.)

This also speeds up startup by about 5% in the hot-cache case, and 12%
in the cold-cache case: not terribly significant, perhaps. (This speedup
happens because we no longer have to read in modules.dep or do a
filesystem walk to find modules unless a module is requested that is
not in the CTF archive.)

Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Orabug: 25815372