dtrace: normalization of addr in the mod() does not work
The normalization of addr must also check if the address belongs
to the module's data ranges. It is required to pick a single value
that represents the module otherwise we will aggregate same key
into multiple buckets. The new value is picked either from first
text range (or data in case text is missing). In other cases the
addr is not normalized.

Orabug: 26826564

Signed-off-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>