Skip tst.system-daemon.sh on the release branch.
We can't really reliably test what this is trying to test until we can
at least reduce the number of garbage stack frames ustack() returns.
Maybe that doesn't mean DWARF, maybe it just means a bit of sanity
checking -- but it won't do right now, when it fails spuriously
on about 20% of runs.