ip provider ipv6_tclass and ipv6_flow are not set correctly
The ipv6_tclass and ipv6_flow fields of the ipv6info_t structure
defined by the ip provider were assigned from the IPv6 header without
conversion to host order, which caused these values to be set
incorrectly. This fix consists of adding the necessary network-to-host
conversions when computing these fields.

Signed-off-by: Nicolas Droux <nicolas.droux@oracle.com>
Reviewed-by: Alan Maguire <alan.maguire@oracle.com>