Test retrieval of args 5th and up for FBT entry probes
This commit introduces the tst.entryargs.d and tst.manyargs.d tests
that exercise the argument retrieval mechanism in FBT for arguments
arg5 and beyond.  The underlying mechanism for the test was modified,
which makes a change to the tst.args-sdt.d test (and results) needed
also.

New test: tst.entryargs.d - compares arguments between FBT and syscall
New test: tst.manyargs.d - compares arguments between FBT and SDT

The tst.basic.d FBT test was also modified because it was apparently
failing to rint arg2 in the list of arguments.  This is not a functional
difference but was a silly omission.

Orabug: 27282243
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>